Warnings fixes (mostly MSVC)
When compiling with MSVC, `/W4` is already enabled.
This is a very comprehensive set of warnings,
and a lot of them are turned into errors when combined with the `/WX` flag.

These flags are enabled globally,
so this patch disables some of the MSVC warnings for external libraries
and tries to fix errors in spirv2clc.

To aim for parity with MSVC, this patch enables `-Werror` and a few extra warnings for GNU style compilers.
The main missing warning was `-Wshadow`.
`-Wstrict-aliasing` had to be partially disabled using pragmas
in order to allow some `reinterpret_cast` calls.